Upgrade go-control-plane
Move the function to compare the equality of slices of envoy resources to the "envoy" packaage.
Use this function in the tests to assess equality due to the fact that comparing protobuf structs
with DeepEqual can have unexpected results.
